+++ lib/bigcommerce/resource.rb
@@ -10,20 +10,24 @@ module Bigcommerce
 
     private
 
+    def api_version
+      :v2
+    end
+
     def get(path, params = {})
-      connection.get(path, params)
+      connection.get(path, params, api_version: api_version)
     end
 
     def post(path, body = {})
-      connection.post(path, body)
+      connection.post(path, body, api_version: api_version)
     end
 
     def put(path, body = {})
-      connection.put(path, body)
+      connection.put(path, body, api_version: api_version)
     end
 
     def delete(path, params = {})
-      connection.delete(path, params)
+      connection.delete(path, params, api_version: api_version)
     end
   end
 end